Problem set 1 (#1)
* Add files via upload

* Delete Problem1.c

* Add files via upload

* Add files via upload

a simple calculator for plus , minus, multiplication , division with swith statements.

* Add files via upload

shows if a year is a leap year or not.
the input  year is a leap year if it is perfectly divisible by 400.
Or the year is a leap year if it is perfectly divisible by 4 but not by 100.
All other years are not leap years.

* Add files via upload

input a cm value and the code will turn it into meter and kilometer

* Add files via upload

user will input Principle , time and rate. And this code will show the simple interest of this

* Add files via upload

user will input two numbers and this code will show the summation of all int numbers between those two numbers

* table

input a number and this code will show you the multiplication tabl of it

* Even or Odd

this is a example of switch topic. in this code the user will input a number and this code will show if its even or odd or zero. iv used switch for this code and also if else for zero.

* circle.c

a program where user will input radius of a circle and the code will output diameter, circumference and area of the circle.

* swapping

it just swaps two numbers.